renaming templates
